Changed question_id to UIInt instead of Double as it actually is an unsigned integer.

+++ StackApp/StackApp/Question.swift
@@ -11,5 +11,5 @@ import Foundation
 struct Question {
     var title: String?
     var last_activity_date: Double?
-    var question_id: Double?
+    var question_id: UInt?
 }
